% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{Array}
\alias{Array}
\title{Array type}
\usage{
Array(items)
}
\arguments{
\item{items}{The type that items in the array should be}
}
\value{
A \code{list} of class \code{Array} describing the valid \code{items} of an array
}
\description{
Array type
}
\examples{
Array(numeric)
Array("numeric")
Array(Person)
Array(Union(character, Person))
}
