% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{as_scalar}
\alias{as_scalar}
\title{Declare that a node is scalar}
\usage{
as_scalar(node)
}
\arguments{
\item{node}{The node to declare as a scalar}
}
\value{
The node with the \code{scalar} class added to it
}
\description{
Declare that a node is scalar
}
