% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{as_type}
\alias{as_type}
\title{Coerce a value to conform to the type
Principally, marks values as scalar where possible}
\usage{
as_type(value, type)
}
\arguments{
\item{value}{The value to coerce}

\item{type}{The type to coerce it to}
}
\value{
The value, possibly coerced to the type
}
\description{
Coerce a value to conform to the type
Principally, marks values as scalar where possible
}
