% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{BlockContent}
\alias{BlockContent}
\title{Union type for valid block content.}
\format{
An object of class \code{Union} of length 1.
}
\usage{
BlockContent
}
\value{
A \code{list} of class \code{Union} describing valid subtypes of this type
}
\description{
Union type for valid block content.
}
\keyword{datasets}
