% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{BooleanValidator}
\alias{BooleanValidator}
\title{A schema specifying that a node must be a boolean value.}
\usage{
BooleanValidator(id, meta)
}
\arguments{
\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{BooleanValidator}
}
\description{
A schema specifying that a node must be a boolean value.
}
\seealso{
\code{\link{Entity}}
}
