% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Brand}
\alias{Brand}
\title{A brand used by an organization or person for labeling a product, product group, or similar.}
\usage{
Brand(
  name,
  alternateNames,
  description,
  id,
  identifiers,
  images,
  logo,
  meta,
  reviews,
  url
)
}
\arguments{
\item{name}{The name of the item. \bold{Required}.}

\item{alternateNames}{Alternate names (aliases) for the item.}

\item{description}{A description of the item.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{logo}{A logo associated with the brand.}

\item{meta}{Metadata associated with this item.}

\item{reviews}{Reviews of the brand.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{Brand}
}
\description{
A brand used by an organization or person for labeling a product, product group, or similar.
}
\seealso{
\code{\link{Thing}}
}
