% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{check_property}
\alias{check_property}
\title{Check that a value is present if required and conforms to the
specified type for a property.}
\usage{
check_property(type_name, property_name, is_required, is_missing, type, value)
}
\arguments{
\item{type_name}{The name of the type that they property is on}

\item{property_name}{The name of the property}

\item{is_required}{Is a value for the property required?}

\item{is_missing}{Is a value for the property missing?}

\item{type}{The type to check against}

\item{value}{The value to check}
}
\value{
The value, if it is valid, nothing otherwise
}
\description{
Check that a value is present if required and conforms to the
specified type for a property.
}
