% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{CiteGroup}
\alias{CiteGroup}
\title{A group of \code{Cite} nodes}
\usage{
CiteGroup(items, id, meta)
}
\arguments{
\item{items}{One or more \code{Cite}s to be referenced in the same surrounding text. \bold{Required}.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{CiteGroup}
}
\description{
A group of \code{Cite} nodes
}
\seealso{
\code{\link{Entity}}
}
