% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Code}
\alias{Code}
\title{Base type for code nodes e.g. \code{CodeBlock}, \code{CodeExpression}.}
\usage{
Code(text, format, id, meta, programmingLanguage)
}
\arguments{
\item{text}{The text of the code. \bold{Required}.}

\item{format}{Media type, typically expressed using a MIME format, of the code.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}

\item{programmingLanguage}{The programming language of the code.}
}
\value{
A \code{list} of class \code{Code}
}
\description{
Base type for code nodes e.g. \code{CodeBlock}, \code{CodeExpression}.
}
\seealso{
\code{\link{Entity}}
}
