% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{CodeBlock}
\alias{CodeBlock}
\title{A code block.}
\usage{
CodeBlock(text, exportFrom, format, id, importTo, meta, programmingLanguage)
}
\arguments{
\item{text}{The text of the code. \bold{Required}.}

\item{exportFrom}{A compilation directive giving the name of the variable to export into the content of the code block.}

\item{format}{Media type, typically expressed using a MIME format, of the code.}

\item{id}{The identifier for this item.}

\item{importTo}{A compilation directive giving the name of the variable to import the content of the code block as.}

\item{meta}{Metadata associated with this item.}

\item{programmingLanguage}{The programming language of the code.}
}
\value{
A \code{list} of class \code{CodeBlock}
}
\description{
A code block.
}
\seealso{
\code{\link{Code}}
}
