% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{CodeExpression}
\alias{CodeExpression}
\title{An expression defined in programming language source code.}
\usage{
CodeExpression(text, errors, format, id, meta, output, programmingLanguage)
}
\arguments{
\item{text}{The text of the code. \bold{Required}.}

\item{errors}{Errors when compiling or executing the chunk.}

\item{format}{Media type, typically expressed using a MIME format, of the code.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}

\item{output}{The value of the expression when it was last evaluated.}

\item{programmingLanguage}{The programming language of the code.}
}
\value{
A \code{list} of class \code{CodeExpression}
}
\description{
An expression defined in programming language source code.
}
\seealso{
\code{\link{CodeFragment}}
}
