% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{CodeFragment}
\alias{CodeFragment}
\title{Inline code.}
\usage{
CodeFragment(text, format, id, meta, programmingLanguage)
}
\arguments{
\item{text}{The text of the code. \bold{Required}.}

\item{format}{Media type, typically expressed using a MIME format, of the code.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}

\item{programmingLanguage}{The programming language of the code.}
}
\value{
A \code{list} of class \code{CodeFragment}
}
\description{
Inline code.
}
\seealso{
\code{\link{Code}}
}
