% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ConstantValidator}
\alias{ConstantValidator}
\title{A validator specifying a constant value that a node must have.}
\usage{
ConstantValidator(id, meta, value)
}
\arguments{
\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}

\item{value}{The value that the node must have.}
}
\value{
A \code{list} of class \code{ConstantValidator}
}
\description{
A validator specifying a constant value that a node must have.
}
\seealso{
\code{\link{Entity}}
}
