% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ContactPoint}
\alias{ContactPoint}
\title{A contact point, for example, a R&D department.}
\usage{
ContactPoint(
  alternateNames,
  availableLanguages,
  description,
  emails,
  id,
  identifiers,
  images,
  meta,
  name,
  telephoneNumbers,
  url
)
}
\arguments{
\item{alternateNames}{Alternate names (aliases) for the item.}

\item{availableLanguages}{Languages (human not programming) in which it is possible to communicate with the organization/department etc.}

\item{description}{A description of the item.}

\item{emails}{Email address for correspondence.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the item.}

\item{telephoneNumbers}{Telephone numbers for the contact point.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{ContactPoint}
}
\description{
A contact point, for example, a R&D department.
}
\seealso{
\code{\link{Thing}}
}
