% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{DatatableColumn}
\alias{DatatableColumn}
\title{A column of data within a Datatable.}
\usage{
DatatableColumn(
  name,
  values,
  alternateNames,
  description,
  id,
  identifiers,
  images,
  meta,
  url,
  validator
)
}
\arguments{
\item{name}{The name of the item. \bold{Required}.}

\item{values}{The data values of the column. \bold{Required}.}

\item{alternateNames}{Alternate names (aliases) for the item.}

\item{description}{A description of the item.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{url}{The URL of the item.}

\item{validator}{The validator to use to validate data in the column.}
}
\value{
A \code{list} of class \code{DatatableColumn}
}
\description{
A column of data within a Datatable.
}
\seealso{
\code{\link{Thing}}
}
