% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Date}
\alias{Date}
\title{A date encoded as a ISO 8601 string.}
\usage{
Date(value, id, meta)
}
\arguments{
\item{value}{The date as an ISO 8601 string. \bold{Required}.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{Date}
}
\description{
A date encoded as a ISO 8601 string.
}
\seealso{
\code{\link{Entity}}
}
