% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{DefinedTerm}
\alias{DefinedTerm}
\title{A word, name, acronym, phrase, etc. with a formal definition.}
\usage{
DefinedTerm(
  name,
  alternateNames,
  description,
  id,
  identifiers,
  images,
  meta,
  termCode,
  url
)
}
\arguments{
\item{name}{The name of the item. \bold{Required}.}

\item{alternateNames}{Alternate names (aliases) for the item.}

\item{description}{A description of the item.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{termCode}{A code that identifies this DefinedTerm within a DefinedTermSet}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{DefinedTerm}
}
\description{
A word, name, acronym, phrase, etc. with a formal definition.
}
\seealso{
\code{\link{Thing}}
}
