% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Entity}
\alias{Entity}
\title{The most simple compound (ie. non-atomic like \code{number}, \code{string} etc) type.}
\usage{
Entity(id, meta)
}
\arguments{
\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{Entity}
}
\description{
The most simple compound (ie. non-atomic like \code{number}, \code{string} etc) type.
}
