% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{Enum}
\alias{Enum}
\title{An enumeration}
\usage{
Enum(...)
}
\arguments{
\item{...}{The values in the enumeration}
}
\value{
A \code{list} of class \code{Enum} describing the valid \code{values} in an enumeration
}
\description{
An enumeration
}
\examples{
Enum("apple", "pear", "orange")
}
