% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Function}
\alias{Function}
\title{A function with a name, which might take Parameters and return a value of a certain type.}
\usage{
Function(id, meta, name, parameters, returns)
}
\arguments{
\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the function.}

\item{parameters}{The parameters of the function.}

\item{returns}{The return type of the function.}
}
\value{
A \code{list} of class \code{Function}
}
\description{
A function with a name, which might take Parameters and return a value of a certain type.
}
\seealso{
\code{\link{Entity}}
}
