% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Heading}
\alias{Heading}
\title{Heading}
\usage{
Heading(content, depth, id, meta)
}
\arguments{
\item{content}{Content of the heading. \bold{Required}.}

\item{depth}{The depth of the heading.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{Heading}
}
\description{
Heading
}
\seealso{
\code{\link{Entity}}
}
