% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Include}
\alias{Include}
\title{A directive to include content from an external source (e.g. file, URL) or content.}
\usage{
Include(source, content, format, id, meta)
}
\arguments{
\item{source}{The source of the content, a URL or file path, or the content itself. \bold{Required}.}

\item{content}{The content to be included.}

\item{format}{Media type, typically expressed using a MIME format, of the source content.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{Include}
}
\description{
A directive to include content from an external source (e.g. file, URL) or content.
}
\seealso{
\code{\link{Entity}}
}
