% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{InlineContent}
\alias{InlineContent}
\title{Union type for valid inline content.}
\format{
An object of class \code{Union} of length 1.
}
\usage{
InlineContent
}
\value{
A \code{list} of class \code{Union} describing valid subtypes of this type
}
\description{
Union type for valid inline content.
}
\keyword{datasets}
