% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{is_class}
\alias{is_class}
\title{Is a value of a particular class}
\usage{
is_class(value, clas)
}
\arguments{
\item{value}{The value to check}

\item{clas}{The class the check against}
}
\value{
TRUE if the value has the class as its last class, FALSE otherwise
}
\description{
Is a value of a particular class
}
