% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{is_type}
\alias{is_type}
\title{Does a value conform to the type?}
\usage{
is_type(value, type)
}
\arguments{
\item{value}{The value to check}

\item{type}{The type to check against}
}
\value{
TRUE if the value conforms to the type, FALSE otherwise
}
\description{
Does a value conform to the type?
}
