% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{last_class}
\alias{last_class}
\title{Get the last class for an object
The last class is usually the "highest" in the inheritance tree}
\usage{
last_class(obj)
}
\arguments{
\item{obj}{The object to get the last class for}
}
\value{
The class that is last in the list of classes for the object
}
\description{
Get the last class for an object
The last class is usually the "highest" in the inheritance tree
}
