% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Link}
\alias{Link}
\title{A hyperlink to other pages, sections within the same document, resources, or any URL.}
\usage{
Link(content, target, exportFrom, id, importTo, meta, relation, title)
}
\arguments{
\item{content}{The textual content of the link. \bold{Required}.}

\item{target}{The target of the link. \bold{Required}.}

\item{exportFrom}{A compilation directive giving the name of the variable to export to the link target.}

\item{id}{The identifier for this item.}

\item{importTo}{A compilation directive giving the name of the variable to import the link target as.}

\item{meta}{Metadata associated with this item.}

\item{relation}{The relation between the target and the current thing.}

\item{title}{A title for the link.}
}
\value{
A \code{list} of class \code{Link}
}
\description{
A hyperlink to other pages, sections within the same document, resources, or any URL.
}
\seealso{
\code{\link{Entity}}
}
