% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{List}
\alias{List}
\title{A list of items.}
\usage{
List(items, id, meta, order)
}
\arguments{
\item{items}{The items in the list \bold{Required}.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}

\item{order}{Type of ordering.}
}
\value{
A \code{list} of class \code{List}
}
\description{
A list of items.
}
\seealso{
\code{\link{Entity}}
}
