% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ListItem}
\alias{ListItem}
\title{A single item in a list.}
\usage{
ListItem(
  alternateNames,
  content,
  description,
  id,
  identifiers,
  images,
  isChecked,
  item,
  meta,
  name,
  position,
  url
)
}
\arguments{
\item{alternateNames}{Alternate names (aliases) for the item.}

\item{content}{The content of the list item.}

\item{description}{A description of the item.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{isChecked}{A flag to indicate if this list item is checked.}

\item{item}{The item represented by this list item.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the item.}

\item{position}{The position of the item in a series or sequence of items.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{ListItem}
}
\description{
A single item in a list.
}
\seealso{
\code{\link{Thing}}
}
