% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Mark}
\alias{Mark}
\title{A base class for nodes that mark some other inline content in some way (e.g. as being emphasised, or quoted).}
\usage{
Mark(content, id, meta)
}
\arguments{
\item{content}{The content that is marked. \bold{Required}.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{Mark}
}
\description{
A base class for nodes that mark some other inline content in some way (e.g. as being emphasised, or quoted).
}
\seealso{
\code{\link{Entity}}
}
