% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{MarkTypes}
\alias{MarkTypes}
\title{All type schemas that are derived from Mark}
\format{
An object of class \code{Union} of length 1.
}
\usage{
MarkTypes
}
\value{
A \code{list} of class \code{Union} describing valid subtypes of this type
}
\description{
All type schemas that are derived from Mark
}
\keyword{datasets}
