% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{MathFragment}
\alias{MathFragment}
\title{A fragment of math, e.g a variable name, to be treated as inline content.}
\usage{
MathFragment(text, errors, id, mathLanguage, meta)
}
\arguments{
\item{text}{The text of the equation in the language. \bold{Required}.}

\item{errors}{Errors that occurred when parsing the math equation.}

\item{id}{The identifier for this item.}

\item{mathLanguage}{The language used for the equation e.g tex, mathml, asciimath.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{MathFragment}
}
\description{
A fragment of math, e.g a variable name, to be treated as inline content.
}
\seealso{
\code{\link{Math}}
}
