% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{MathTypes}
\alias{MathTypes}
\title{All type schemas that are derived from Math}
\format{
An object of class \code{Union} of length 1.
}
\usage{
MathTypes
}
\value{
A \code{list} of class \code{Union} describing valid subtypes of this type
}
\description{
All type schemas that are derived from Math
}
\keyword{datasets}
