% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{MediaObjectTypes}
\alias{MediaObjectTypes}
\title{All type schemas that are derived from MediaObject}
\format{
An object of class \code{Union} of length 1.
}
\usage{
MediaObjectTypes
}
\value{
A \code{list} of class \code{Union} describing valid subtypes of this type
}
\description{
All type schemas that are derived from MediaObject
}
\keyword{datasets}
