% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{MonetaryGrant}
\alias{MonetaryGrant}
\title{A monetary grant.}
\usage{
MonetaryGrant(
  alternateNames,
  amounts,
  description,
  fundedItems,
  funders,
  id,
  identifiers,
  images,
  meta,
  name,
  sponsors,
  url
)
}
\arguments{
\item{alternateNames}{Alternate names (aliases) for the item.}

\item{amounts}{The amount of money.}

\item{description}{A description of the item.}

\item{fundedItems}{Indicates an item funded or sponsored through a Grant.}

\item{funders}{A person or organization that supports (sponsors) something through some kind of financial contribution.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the item.}

\item{sponsors}{A person or organization that supports a thing through a pledge, promise, or financial contribution.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{MonetaryGrant}
}
\description{
A monetary grant.
}
\seealso{
\code{\link{Grant}}
}
