% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{Node}
\alias{Node}
\title{Union type for all valid nodes.}
\format{
An object of class \code{Union} of length 1.
}
\usage{
Node
}
\value{
A \code{list} of class \code{Union} describing valid subtypes of this type
}
\description{
Union type for all valid nodes.
}
\keyword{datasets}
