% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{NontextualAnnotation}
\alias{NontextualAnnotation}
\title{Inline text that has a non-textual annotation.}
\usage{
NontextualAnnotation(content, id, meta)
}
\arguments{
\item{content}{The content that is marked. \bold{Required}.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{NontextualAnnotation}
}
\description{
Inline text that has a non-textual annotation.
}
\seealso{
\code{\link{Mark}}
}
