% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{NumberValidatorTypes}
\alias{NumberValidatorTypes}
\title{All type schemas that are derived from NumberValidator}
\format{
An object of class \code{Union} of length 1.
}
\usage{
NumberValidatorTypes
}
\value{
A \code{list} of class \code{Union} describing valid subtypes of this type
}
\description{
All type schemas that are derived from NumberValidator
}
\keyword{datasets}
