% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Paragraph}
\alias{Paragraph}
\title{Paragraph}
\usage{
Paragraph(content, id, meta)
}
\arguments{
\item{content}{The contents of the paragraph. \bold{Required}.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{Paragraph}
}
\description{
Paragraph
}
\seealso{
\code{\link{Entity}}
}
