% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PostalAddress}
\alias{PostalAddress}
\title{A physical mailing address.}
\usage{
PostalAddress(
  addressCountry,
  addressLocality,
  addressRegion,
  alternateNames,
  availableLanguages,
  description,
  emails,
  id,
  identifiers,
  images,
  meta,
  name,
  postOfficeBoxNumber,
  postalCode,
  streetAddress,
  telephoneNumbers,
  url
)
}
\arguments{
\item{addressCountry}{The country.}

\item{addressLocality}{The locality in which the street address is, and which is in the region.}

\item{addressRegion}{The region in which the locality is, and which is in the country.}

\item{alternateNames}{Alternate names (aliases) for the item.}

\item{availableLanguages}{Languages (human not programming) in which it is possible to communicate with the organization/department etc.}

\item{description}{A description of the item.}

\item{emails}{Email address for correspondence.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the item.}

\item{postOfficeBoxNumber}{The post office box number.}

\item{postalCode}{The postal code.}

\item{streetAddress}{The street address.}

\item{telephoneNumbers}{Telephone numbers for the contact point.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{PostalAddress}
}
\description{
A physical mailing address.
}
\seealso{
\code{\link{ContactPoint}}
}
