% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Product}
\alias{Product}
\title{Any offered product or service. For example, a pair of shoes; a haircut; or an episode of a TV show streamed online.}
\usage{
Product(
  alternateNames,
  brands,
  description,
  id,
  identifiers,
  images,
  logo,
  meta,
  name,
  productID,
  url
)
}
\arguments{
\item{alternateNames}{Alternate names (aliases) for the item.}

\item{brands}{Brands that the product is labelled with.}

\item{description}{A description of the item.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{logo}{The logo of the product.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the item.}

\item{productID}{Product identification code.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{Product}
}
\description{
Any offered product or service. For example, a pair of shoes; a haircut; or an episode of a TV show streamed online.
}
\seealso{
\code{\link{Thing}}
}
