% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{PropertyValue}
\alias{PropertyValue}
\title{A property-value pair.}
\usage{
PropertyValue(
  value,
  alternateNames,
  description,
  id,
  identifiers,
  images,
  meta,
  name,
  propertyID,
  url
)
}
\arguments{
\item{value}{The value of the property. \bold{Required}.}

\item{alternateNames}{Alternate names (aliases) for the item.}

\item{description}{A description of the item.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the item.}

\item{propertyID}{A commonly used identifier for the characteristic represented by the property.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{PropertyValue}
}
\description{
A property-value pair.
}
\seealso{
\code{\link{Thing}}
}
