% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Quote}
\alias{Quote}
\title{Inline, quoted content.}
\usage{
Quote(content, cite, id, meta)
}
\arguments{
\item{content}{The content that is marked. \bold{Required}.}

\item{cite}{The source of the quote.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{Quote}
}
\description{
Inline, quoted content.
}
\seealso{
\code{\link{Mark}}
}
