% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{QuoteBlock}
\alias{QuoteBlock}
\title{A section quoted from somewhere else.}
\usage{
QuoteBlock(content, cite, id, meta)
}
\arguments{
\item{content}{The content of the quote. \bold{Required}.}

\item{cite}{The source of the quote.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{QuoteBlock}
}
\description{
A section quoted from somewhere else.
}
\seealso{
\code{\link{Entity}}
}
