% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SoftwareEnvironment}
\alias{SoftwareEnvironment}
\title{A computational environment.}
\usage{
SoftwareEnvironment(
  name,
  adds,
  alternateNames,
  description,
  extends,
  id,
  identifiers,
  images,
  meta,
  removes,
  url
)
}
\arguments{
\item{name}{The name of the item. \bold{Required}.}

\item{adds}{The packages that this environment adds to the base environments listed under \code{extends} (if any).,}

\item{alternateNames}{Alternate names (aliases) for the item.}

\item{description}{A description of the item.}

\item{extends}{Other environments that this environment extends by adding or removing packages.,}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{removes}{The packages that this environment removes from the base environments listed under \code{extends} (if any).,}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{SoftwareEnvironment}
}
\description{
A computational environment.
}
\seealso{
\code{\link{Thing}}
}
