% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{SoftwareSourceCode}
\alias{SoftwareSourceCode}
\title{Computer programming source code. Example: Full (compile ready) solutions, code snippet samples, scripts, templates.}
\usage{
SoftwareSourceCode(
  about,
  alternateNames,
  authors,
  codeRepository,
  codeSampleType,
  comments,
  content,
  dateAccepted,
  dateCreated,
  dateModified,
  datePublished,
  dateReceived,
  description,
  editors,
  fundedBy,
  funders,
  genre,
  id,
  identifiers,
  images,
  isPartOf,
  keywords,
  licenses,
  maintainers,
  meta,
  name,
  parts,
  programmingLanguage,
  publisher,
  references,
  runtimePlatform,
  softwareRequirements,
  targetProducts,
  text,
  title,
  url,
  version
)
}
\arguments{
\item{about}{The subject matter of the content.}

\item{alternateNames}{Alternate names (aliases) for the item.}

\item{authors}{The authors of this creative work.}

\item{codeRepository}{Link to the repository where the un-compiled, human readable code and related code is located.}

\item{codeSampleType}{What type of code sample: full (compile ready) solution, code snippet, inline code, scripts, template.}

\item{comments}{Comments about this creative work.}

\item{content}{The structured content of this creative work c.f. property \code{text}.}

\item{dateAccepted}{Date/time of acceptance.}

\item{dateCreated}{Date/time of creation.}

\item{dateModified}{Date/time of most recent modification.}

\item{datePublished}{Date of first publication.}

\item{dateReceived}{Date/time that work was received.}

\item{description}{A description of the item.}

\item{editors}{People who edited the \code{CreativeWork}.}

\item{fundedBy}{Grants that funded the \code{CreativeWork}; reverse of \code{fundedItems}.}

\item{funders}{People or organizations that funded the \code{CreativeWork}.}

\item{genre}{Genre of the creative work, broadcast channel or group.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{isPartOf}{An item or other CreativeWork that this CreativeWork is a part of.}

\item{keywords}{Keywords or tags used to describe this content. Multiple entries in a keywords list are typically delimited by commas.}

\item{licenses}{License documents that applies to this content, typically indicated by URL.}

\item{maintainers}{The people or organizations who maintain this CreativeWork.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the item.}

\item{parts}{Elements of the collection which can be a variety of different elements, such as Articles, Datatables, Tables and more.}

\item{programmingLanguage}{The computer programming language.}

\item{publisher}{A publisher of the CreativeWork.}

\item{references}{References to other creative works, such as another publication, web page, scholarly article, etc.}

\item{runtimePlatform}{Runtime platform or script interpreter dependencies (Example - Java v1, Python2.3, .Net Framework 3.0).}

\item{softwareRequirements}{Dependency requirements for the software.}

\item{targetProducts}{Target operating system or product to which the code applies.}

\item{text}{The textual content of this creative work.}

\item{title}{The title of the creative work.}

\item{url}{The URL of the item.}

\item{version}{The version of the creative work.}
}
\value{
A \code{list} of class \code{SoftwareSourceCode}
}
\description{
Computer programming source code. Example: Full (compile ready) solutions, code snippet samples, scripts, templates.
}
\seealso{
\code{\link{CreativeWork}}
}
