% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{StringValidator}
\alias{StringValidator}
\title{A schema specifying constraints on a string node.}
\usage{
StringValidator(id, maxLength, meta, minLength, pattern)
}
\arguments{
\item{id}{The identifier for this item.}

\item{maxLength}{The maximum length for a string node.}

\item{meta}{Metadata associated with this item.}

\item{minLength}{The minimum length for a string node.}

\item{pattern}{A regular expression that a string node must match.}
}
\value{
A \code{list} of class \code{StringValidator}
}
\description{
A schema specifying constraints on a string node.
}
\seealso{
\code{\link{Entity}}
}
