% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Subscript}
\alias{Subscript}
\title{Subscripted content.}
\usage{
Subscript(content, id, meta)
}
\arguments{
\item{content}{The content that is marked. \bold{Required}.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{Subscript}
}
\description{
Subscripted content.
}
\seealso{
\code{\link{Mark}}
}
