% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TableCell}
\alias{TableCell}
\title{A cell within a \code{Table}.}
\usage{
TableCell(content, cellType, colspan, id, meta, name, rowspan)
}
\arguments{
\item{content}{Contents of the table cell. \bold{Required}.}

\item{cellType}{Indicates whether the cell is a header or data.}

\item{colspan}{How many columns the cell extends.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the cell.}

\item{rowspan}{How many columns the cell extends.}
}
\value{
A \code{list} of class \code{TableCell}
}
\description{
A cell within a \code{Table}.
}
\seealso{
\code{\link{Entity}}
}
