% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TableRow}
\alias{TableRow}
\title{A row within a Table.}
\usage{
TableRow(cells, id, meta, rowType)
}
\arguments{
\item{cells}{An array of cells in the row. \bold{Required}.}

\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}

\item{rowType}{If present, indicates that all cells in this row should be treated as header cells.}
}
\value{
A \code{list} of class \code{TableRow}
}
\description{
A row within a Table.
}
\seealso{
\code{\link{Entity}}
}
