% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{ThematicBreak}
\alias{ThematicBreak}
\title{A thematic break, such as a scene change in a story, a transition to another topic, or a new document.}
\usage{
ThematicBreak(id, meta)
}
\arguments{
\item{id}{The identifier for this item.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{ThematicBreak}
}
\description{
A thematic break, such as a scene change in a story, a transition to another topic, or a new document.
}
\seealso{
\code{\link{Entity}}
}
