% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Thing}
\alias{Thing}
\title{The most generic type of item.}
\usage{
Thing(alternateNames, description, id, identifiers, images, meta, name, url)
}
\arguments{
\item{alternateNames}{Alternate names (aliases) for the item.}

\item{description}{A description of the item.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{name}{The name of the item.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{Thing}
}
\description{
The most generic type of item.
}
\seealso{
\code{\link{Entity}}
}
