% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{TupleValidator}
\alias{TupleValidator}
\title{A validator specifying constraints on an array of heterogeneous items.}
\usage{
TupleValidator(id, items, meta)
}
\arguments{
\item{id}{The identifier for this item.}

\item{items}{An array of validators specifying the constraints on each successive item in the array.}

\item{meta}{Metadata associated with this item.}
}
\value{
A \code{list} of class \code{TupleValidator}
}
\description{
A validator specifying constraints on an array of heterogeneous items.
}
\seealso{
\code{\link{Entity}}
}
