% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typing.R
\name{Union}
\alias{Union}
\title{Union type}
\usage{
Union(...)
}
\arguments{
\item{...}{The types in the union}
}
\value{
A \code{list} of class \code{Union} describing the valid sub \code{types} of a union type
}
\description{
Union type
}
\examples{
Union(character, Person)
}
