% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{ValidatorTypes}
\alias{ValidatorTypes}
\title{Union type for all validator types.}
\format{
An object of class \code{Union} of length 1.
}
\usage{
ValidatorTypes
}
\value{
A \code{list} of class \code{Union} describing valid subtypes of this type
}
\description{
Union type for all validator types.
}
\keyword{datasets}
