% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Variable}
\alias{Variable}
\title{A variable representing a name / value pair.}
\usage{
Variable(name, id, isReadonly, meta, validator, value)
}
\arguments{
\item{name}{The name of the variable. \bold{Required}.}

\item{id}{The identifier for this item.}

\item{isReadonly}{Whether or not a property is mutable. Default is false.}

\item{meta}{Metadata associated with this item.}

\item{validator}{The validator that the value is validated against.}

\item{value}{The value of the variable.}
}
\value{
A \code{list} of class \code{Variable}
}
\description{
A variable representing a name / value pair.
}
\seealso{
\code{\link{Entity}}
}
