% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{VolumeMount}
\alias{VolumeMount}
\title{Describes a volume mount from a host to container.}
\usage{
VolumeMount(
  mountDestination,
  alternateNames,
  description,
  id,
  identifiers,
  images,
  meta,
  mountOptions,
  mountSource,
  mountType,
  name,
  url
)
}
\arguments{
\item{mountDestination}{The mount location inside the container. \bold{Required}.}

\item{alternateNames}{Alternate names (aliases) for the item.}

\item{description}{A description of the item.}

\item{id}{The identifier for this item.}

\item{identifiers}{Any kind of identifier for any kind of Thing.}

\item{images}{Images of the item.}

\item{meta}{Metadata associated with this item.}

\item{mountOptions}{A list of options to use when applying the mount.}

\item{mountSource}{The mount source directory on the host.}

\item{mountType}{The type of mount.}

\item{name}{The name of the item.}

\item{url}{The URL of the item.}
}
\value{
A \code{list} of class \code{VolumeMount}
}
\description{
Describes a volume mount from a host to container.
}
\seealso{
\code{\link{Thing}}
}
